chore(email): move header types to their own file
