+++ packages/email/src/message.ts
@@ -1,6 +1,5 @@
-import { type Many, many } from '@seedcompany/common';
-import type { PathLike } from 'node:fs';
-import type { Readable } from 'node:stream';
+import { many } from '@seedcompany/common';
+import type { MessageHeaders } from './headers.type.js';
 
 export class EmailMessage {
   readonly templateName: string;
@@ -28,56 +27,3 @@ export class EmailMessage {
     });
   }
 }
-
-// Below is inlined from the `emailjs` library to avoid using their loose source files
-// I changed to immutable
-
-type MessageHeaders = Readonly<{
-  [index: string]:
-    | boolean
-    | Many<string>
-    | null
-    | undefined
-    | Many<MessageAttachment>;
-  'content-type'?: string;
-  'message-id'?: string;
-  'return-path'?: string | null;
-  date?: string;
-  from: Many<string>;
-  to: Many<string>;
-  cc?: Many<string>;
-  bcc?: Many<string>;
-  subject: string;
-  text: string | null;
-  attachment?: Many<MessageAttachment>;
-}>;
-
-type MessageAttachment = Readonly<{
-  [index: string]:
-    | string
-    | boolean
-    | Many<MessageAttachment>
-    | MessageAttachmentHeaders
-    | Readable
-    | PathLike
-    | undefined;
-  name?: string;
-  headers?: MessageAttachmentHeaders;
-  inline?: boolean;
-  alternative?: MessageAttachment | boolean;
-  related?: readonly MessageAttachment[];
-  data?: string;
-  encoded?: boolean;
-  stream?: Readable;
-  path?: PathLike;
-  type?: string;
-  charset?: string;
-  method?: string;
-}>;
-
-type MessageAttachmentHeaders = Readonly<{
-  [index: string]: string | undefined;
-  'content-type'?: string;
-  'content-transfer-encoding'?: BufferEncoding | '7bit' | '8bit';
-  'content-disposition'?: string;
-}>;
